contractor invoice 100% done'

+++ db/migrate/20180912135555_create_contractorpos.rb
@@ -3,7 +3,8 @@ class CreateContractorpos < ActiveRecord::Migration[5.2]
     create_table :contractorpos do |t|
       t.string :number
       t.date :date
-      t.integer :amount
+      t.float :amount
+      # t.integer :amount
       t.text :description
       t.belongs_to :contractorquote, foreign_key: true
