+++ db/migrate/20180913055227_create_stages.rb
@@ -1,7 +1,8 @@
 class CreateStages < ActiveRecord::Migration[5.2]
   def change
     create_table :stages do |t|
-      t.integer :percentage
+      t.float :percentage
+      # t.integer :percentage
       t.belongs_to :milestone, foreign_key: true
 
       t.timestamps
